feat: Weight calculations (prototype ready for graph)

+++ pkg/documents/document.go
@@ -2,5 +2,5 @@ package documents
 
 type Document interface {
 	Path() string
-	KeywordWeights() map[string]float32
+	KeywordWeights() map[string]float64
 }
